requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.40.1
google-generativeai==0.3.1
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
idna==3.10
Jinja2==3.1.2
MarkupSafe==3.0.2
passlib==1.7.4
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.0
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.6
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.23.2

# SQLAlchemy 관련 (sessionmaker 오류)
#pip install sqlalchemy

# Google Generative AI 관련
#pip install google-generativeai

# Passlib 관련 (비밀번호 해싱 라이브러리)
#pip install passlib[bcrypt]

# python-jose (JWT 인증에 필요)
#pip install python-jose[cryptography]